mass update

+++ apg-tools/webhook.php
@@ -1,43 +1,40 @@
 <?php
-
-define( 'WP_ADMIN', TRUE );
-define( 'WP_NETWORK_ADMIN', TRUE );
-define( 'WP_USER_ADMIN', TRUE );
-
-require_once('../../../wp-load.php');
-require_once(ABSPATH.'wp-admin/includes/admin.php');
-require_once(ABSPATH.'wp-admin/includes/file.php');
-require_once(ABSPATH.'wp-admin/includes/plugin.php');
-
+ini_set('max_execution_time', 0);
 try {
+  define( 'WP_ADMIN', TRUE );
+  define( 'WP_NETWORK_ADMIN', TRUE );
+  define( 'WP_USER_ADMIN', TRUE );
+  require_once('includes/config.php');
+  require_once(ABSPATH.'wp-admin/includes/admin.php');
+  require_once(ABSPATH.'wp-admin/includes/file.php');
+  require_once(ABSPATH.'wp-admin/includes/plugin.php');
+  require_once(ABSPATH.'wp-content/plugins/apg-tools/includes/helper.php');
+
   $post = $_POST;
-  $plugins = [];
-  // file_put_contents('logs.txt', json_encode($post));die;
-  foreach($post['plugins'] as $key => $value){
-    if($value['is_external_link'] == 0){
-      $file_url = $post['url_download']."/{$value['id']}";
-    }else{
-      $file_url = $value['external_link'];
+
+  if($post['token_auth'] == apg_config()->token_auth){
+    foreach($post['plugins'] as $key => $value){
+      if($value['is_external_link'] == 0){
+        $file_url = $post['url_download']."/{$value['file_name']}";
+      }else{
+        $file_url = $value['external_link'];
+      }
+      $tmp_file = download_url($file_url);
+      copy($tmp_file, $value['file_name']);
+      unlink($tmpfile);
+      // extract plugin to plugins folder
+      $zip = new ZipArchive;
+      $zip->open($value['file_name']);
+      $zip->extractTo(ABSPATH.'wp-content/plugins/');
+      $zip->close();
+      // remove zip file
+      unlink($value['file_name']);
+      // activate plugin
+      apg_activate_plugin($value['slug'].'/'.$value['core_file']);
     }
-    $tmp_file = download_url($file_url);
-    copy($tmp_file, $value['file_name']);
-    unlink($tmpfile);
-    // extract plugin to plugins folder
-    $zip = new ZipArchive;
-    $zip->open($value['file_name']);
-    $zip->extractTo(ABSPATH.'wp-content/plugins/');
-    $zip->close();
-    // remove zip file
-    unlink($value['file_name']);
-    // activate plugin
-    // $res = activate_plugin(ABSPATH.'wp-content/plugins/'.$value['slug'].'/'.$value['core_file']);
-    $plugins[] = ABSPATH.'wp-content/plugins/'.$value['slug'].'/'.$value['core_file'];
-  }
-  // activate plugin
-  $res = activate_plugins($plugins);
-  // get error response if there
-  if(is_wp_error($result)){
-    file_put_contents("logs.txt", $res->get_error_messages());
+    return true;
+  }else{
+    return http_response_code(403);
   }
 } catch (\Throwable $th) {
   file_put_contents('log.txt', $th);
